Move getting deps from import map out to separate utility function
