+++ Cargo.toml
@@ -356,5 +356,5 @@ opt-level = 3
 [patch.crates-io]
 deno_lockfile = { path = "../deno_lockfile" }
 deno_npm = { path = "../deno_npm" }
-deno_semver = { path = "../deno_semver" }
 deno_graph = { path = "../deno_graph" }
+deno_semver = { path = "../deno_semver" }
